#8 works

+++ Cefalo.farhadcodes-a-CP-blog.Service/Services/StoryService.cs
@@ -2,7 +2,10 @@
 using Cefalo.farhadcodes_a_CP_blog.Database.Models;
 using Cefalo.farhadcodes_a_CP_blog.Repository.Contracts;
 using Cefalo.farhadcodes_a_CP_blog.Service.Contracts;
+using Cefalo.farhadcodes_a_CP_blog.Service.CustomExceptions;
 using Cefalo.farhadcodes_a_CP_blog.Service.DTO.Story;
+using Cefalo.farhadcodes_a_CP_blog.Service.Handler.Contracts;
+using Cefalo.farhadcodes_a_CP_blog.Service.Handler.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,10 +18,12 @@ namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
     {
         private readonly IStoryRepository _storyRepository;
         private readonly IMapper _mapper;
-        public StoryService(IStoryRepository storyRepository, IMapper mapper)
+        private readonly IPassword _passwordH;
+        public StoryService(IStoryRepository storyRepository, IMapper mapper, IPassword password)
         {
             _storyRepository = storyRepository;
             _mapper = mapper;
+            _passwordH = password;
         }
 
         public async Task<List<StoryDTO>> GetStories()
@@ -30,6 +35,7 @@ namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
         public async Task<Story> GetStory(int id)
         {
             var story = await _storyRepository.GetStory(id);
+            if (story == null) throw new NotFoundHandler("No Story was found with that Id");
             return story;
         }
 
@@ -44,14 +50,37 @@ namespace Cefalo.farhadcodes_a_CP_blog.Service.Services
 
         public async Task<Story> UpdateStory(int id, UpdateStory body)
         {
-            if (id != body.Id)
-                return null;
+            if (id != body.Id) throw new BadRequestHandler("Bad request!");
+            // no story exists
+            var fetchedStory = await _storyRepository.GetStory(id);
+            if(fetchedStory == null) throw new NotFoundHandler("No Story was found with that Id");
+            // authorization
+            var creationTime = _passwordH.GetTokenCreationTime();
+            if (creationTime == null) throw new UnauthorisedHandler("You're not logged in! Please log in to get access.");
+            DateTime tokenCreationTime = Convert.ToDateTime(creationTime);
+
+            if (DateTime.Compare(tokenCreationTime, fetchedStory.LastModifiedTime) < 0)
+                throw new UnauthorisedHandler("JWT Expired! Login again!");
+            var currUserId = _passwordH.GetLoggedInId();
+            if(currUserId != fetchedStory.AuthorID) throw new ForbiddenHandler("You don't have the permission!");
+            
             Story story = _mapper.Map<Story>(body);
             var updatedStory = await _storyRepository.UpdateStory(id, story);
             return updatedStory;
         }
         public async Task<bool?> DeleteStory(int id)
         {
+            var fetchedStory = await _storyRepository.GetStory(id);
+            if (fetchedStory == null) throw new NotFoundHandler("No Story was found with that Id");
+            var creationTime = _passwordH.GetTokenCreationTime();
+            if (creationTime == null) throw new UnauthorisedHandler("You're not logged in! Please log in to get access.");
+            DateTime tokenCreationTime = Convert.ToDateTime(creationTime);
+
+            if (DateTime.Compare(tokenCreationTime, fetchedStory.LastModifiedTime) < 0)
+                throw new UnauthorisedHandler("JWT Expired! Login again!");
+            var currUserId = _passwordH.GetLoggedInId();
+            if (currUserId != fetchedStory.AuthorID) throw new ForbiddenHandler("You don't have the permission!");
+
             return await _storyRepository.DeleteStory(id);
         }
     }
